requirements/prod.txt
-r common.txt
-r tendenci.txt

# gevent for async workers
gevent==0.13.8

# Memcache
pylibmc==1.2.2
django-pylibmc-sasl==0.2.4

# For Sentry
raven==3.4.0
